# archive/docs/PlatformPortal/Backend_API_Specs.yaml
openapi: 3.1.0
info:
  title: WAOOAW Platform Portal - Backend API Specifications
  description: |
    Complete API specifications for WAOOAW Platform Portal backend integration.
    
    **Purpose**: This document defines all backend APIs required by the Platform Portal 
    to transition from mock data to real platform infrastructure integration.
    
    **Current Status**: 
    - ✅ APIs marked with [IMPLEMENTED] are available in backend
    - ⚠️ APIs marked with [MOCK DATA] return placeholder data
    - ❌ APIs marked with [MISSING] need to be implemented
    
    **Portal Pages**:
    1. Dashboard - Platform metrics and health overview
    2. Queue Monitoring - Task queue metrics and DLQ management  
    3. Workflows - Orchestration monitoring
    4. Agent Factory - Agent lifecycle management
    5. Agent Servicing - Zero-downtime upgrades
    6. Help Desk - Incident tracking and diagnostics
    
    **Base URL**: http://localhost:8000 (development)
    **Authentication**: OAuth 2.0 (GitHub OAuth for portal access)
    
  version: 1.0.0
  contact:
    name: WAOOAW Platform Team
    url: https://github.com/dlai-sd/WAOOAW
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{codespace}-8000.app.github.dev
    description: GitHub Codespaces
    variables:
      codespace:
        default: your-codespace-name
        description: Your GitHub Codespace name
...
paths:
...
  /api/platform/logs:
    get:
      tags: [platform]
      summary: Get System Logs
      description: |
        **Portal Usage**: Logs & Alerts page (Iteration 8 - NOT YET IMPLEMENTED)
        **Status**: [MOCK DATA]
        **Implementation**: backend/app/main.py (stub)
        
        **Real Integration Needs**:
        - Centralized logging system (ELK, Loki, etc.)
        - Log aggregation from all agents
        - Real-time log streaming capability
        
      operationId: getPlatformLogs
      parameters:
        - name: level
          in: query
          description: Filter by log level
          schema:
            type: string
            enum: [DEBUG, INFO, WARN, ERROR]
        - name: agent
          in: query
          description: Filter by agent name
          schema:
            type: string
        - name: search
          in: query
          description: Search in log messages
          schema:
            type: string
        - name: limit
          in: query
          description: Number of logs to return
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: System logs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - timestamp
                    - level
                    - message
                  properties:
                    timestamp:
                      type: string
                      format: date-time
                    level:
                      type: string
                      enum: [DEBUG, INFO, WARN, ERROR]
                    agent:
                      type: string
                    message:
                      type: string
                    trace_id:
                      type: string
                    metadata:
                      type: object
                      additionalProperties: true
...
  /api/platform/alerts/{alert_id}/acknowledge:
    post:
      tags: [platform]
      summary: Acknowledge Alert
      description: |
        **Portal Usage**: Logs & Alerts page - Mark alert as acknowledged
        **Status**: [MOCK DATA]
        
      operationId: acknowledgeAlert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                acknowledged_by:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Alert acknowledged
        '404':
          description: Alert not found

# ============================================================================
# AGENT APIS - Agent Factory & Dashboard Pages
# ============================================================================
...
components:
...
  securitySchemes:
    OAuth2:
      type: oauth2
      description: GitHub OAuth for portal access
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            read:user: Read user profile information
security:
  - OAuth2: [read:user]

# ============================================================================
# IMPLEMENTATION PRIORITY
# ============================================================================
...